Chore: Move db migration to unit conftest
