+++ tests/unit/django/conftest.py
@@ -12,8 +12,8 @@ import ops
 import pytest
 from ops.testing import Harness
 
-from paas_charm.database_migration import DatabaseMigrationStatus
 from examples.django.charm.src.charm import DjangoCharm
+from paas_charm.database_migration import DatabaseMigrationStatus
 
 from .constants import DEFAULT_LAYER, DJANGO_CONTAINER_NAME
